+++ R/c-api.R
@@ -7,13 +7,13 @@ init_c_constants <- function() {
     tilde_thunk_fmls =
       formals(function(...) NULL),
 
-    tilde_thunk_body =
-      expr((!!.Call)(!!(rlang_tilde_eval),
-        sys.call(),
-        "data_mask_arg",
-        "data_mask_top_arg",
-        environment()
-      ))
+    tilde_thunk_body = bquote(.(.Call)(.(rlang_tilde_eval),
+      sys.call(),
+      "data_mask_arg",
+      "data_mask_top_arg",
+      environment(),
+      parent.frame()
+    ))
 
   )
 }